Add script to restart map with new accounts
Creates a new instances table to keep track of restarts so we don't
restart it too often, and adds instance name to the accounts table.

With that in place, we can finally have a script to automatically grab
the completed accounts, write a new accounts file, and restart rocketmap

It's not perfect yet. Need to be able to track multiple runs...
After a few days when blinds happen, we don't want it using old accounts