Fix bug where Faker doesn't always add a symbol
In some edge cases (1-4 in every 100 accounts) passwords get generated
without containing any symbols, causing KinanCore to run in an endless
loop because of the invalid password.

To avoid that we bump symbols to minimum of 2, and if that's not good
enough, we'll keep looping until symbols has at least one value.